@CLAassistant 2019-10-25
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/LeagueSandbox/GameServer?pullRequest=911) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/LeagueSandbox/GameServer?pullRequest=911) before we can accept your contribution.<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/LeagueSandbox/GameServer?pullRequest=911) it.</sub>