@danil179 2019-10-26
I think this indeed optimized a lot of things (and there is no reason to store the scripts in lists). I'll check this and merge if OK.